prompt_test_container/app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.1
numpy==2.2.6
ollama==0.5.1
pandas==2.3.0
pathlib==1.0.1
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scikit-learn==1.7.0
scikit-multilearn==0.2.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
